Fixed the TheDF global causing issues in the TheDF local overwrite of modern cutoff.
Readded the plot option for a 4x3 pdf in the NXN, no success on adding plot_spacers as individual list items sadly.

There is a weird name splicing issue for the Nevil experiment series, fails to name true fashion causing overwrite of the pdfs and double lines appearing in the plot as both Ctrl and PMA locations get added.