Saving these before we get too far down the rabbit hole.
Updated NameCleanup so it handles fixed arguments. Removed SingleColorPlots and replaced with the heading towards modern implementation Utility_Gating plots (a lot of fixing work remains). Updated the manual, and apparently hadn't yet saved the vignette.

Still a bunch of work on Utility_GatingPlots, there is a patchwork for-loop that needs to be cleaned up.