It's raining tSNE implementations. Currently we have Rtsne, which implements the original using an R wrapper for the C++ code, and is used by Catalyst and Condor.
Also brought in openTSNE which is a Python implementation, with no wrapper.

The other alternatives are fltSNE, and opt-SNE which consist of R/python wrappers around C++ and FORTRAN code. Both are not in public active phase unlike the two above. TBD. Worse case scenario for the paper we do all the R side and then just run optSNE in FlowJo as the final hurrah.