Major Changes: Similar to Coereba, moved functions that while useful, not serving purpose for this particular package to OutToPasture package. For luciernaga, only the really old functions, and a couple unmixing tinkering.
A lot of the remaining functions need some tinkering/rewrite.

I also need to decide on whether to leave the dim reduction functions public facing or to internalize them, or just remove the manual pages. I don't want future users dependent on me to do them for their experiments, but yes to use it for the paper questions, etc.