Rewrite of Utility_SingleColorQC is needed now that I have several more months of coding experience under my belt.
Some of the considerations implemented don't hold up with the knowledge of brightness/signature since gained.

And the use of for-loops and suppress warnings is not good practice, so alternate structures and internal function calls to break it up to reasonable size is needed for Bioconductor submission.

It will also allow for easier redirecting when it comes to report generating (beyond current role as an .fcs subcluster generator).