Low on blood glucose, but Concatenate, Downsample and Col_Append seem reachable targets, with options to either return flowframe or send to an .fcs. Need to rewrite after dinner to fix these.
Added a Coereba splitpoint matrix for those 3 specimens, haven't checked the gates though (should NxN it).

The idea of adding Ridge plots (maybe geom_density) to finish the visualization section. Don't feel adding norm algs here is appropiate, but maybe linking to an example script via StepUpCytometry's repo for anyone who wants to adapt for their own use.

We should also consider for standard practice anytime there is a data from subset to have a option that can be overridden in case they want the data returned in raw non-transformed form. The issue with that is is the reversion present in the GS or does it need to be respecified?

And yes, also the other two DimViz via Basilisk, all the QC for Libraries, the references, etc.