Major changes to Utility_Concatinate and Utility_Downsample, updated their importFrom as they were lacking! They now communicate with Utility_ColAppend, and have options for data.frame, flow.frame or .fcs outputs.
For concatinate, we need to replace the ff name parameters as they just copied from original. This will likely need to be a work in progress with improvements on as we go.

Decided to not bother integrating downsample with TSNE or UMAP as it's just two lines of code and they differ from downsample at when they get the data.