I think we are through SingleStainSignatures.
Broad level highlights, we can now subtract AF either using the localized internal AF, or the provisioned average external AF.

We have tracked down the bug (non-rounded values) that was causing low returns from the cluster case when in Cluster Iteration.

And we implemented a SCData = "subtraction" to allow for swapping in and out of raw/subtracted data. Unfortunately, since it's map iterated in, each peak detector is going to have a different internal AF negative. So no way to get out (even with a list at this point).

Unclear what effect returning purified SC signature dependent on the localized AF extract will look like in terms of the raw values.

And yes, a lot of negative values from subtraction.