Up to the concatenated exprs frames for individual fluorophores.
Debating whether to return as summarized data from that depth, or to simply return as a data.frame to the upper layer. The question mark is the Geom_Density plots, which may need individual rows. But having a large amount of exprs extracted sitting at the top would be a lot of RAM compared to the summarized results.

I will pursue the summarized return for now, making compatible for LuciernagaReport(), and we can always tweak the internal functions to return the data when a setting is given (combine with geom density specific function?).