Added the first part of the bead single color handling. From PeakDetectors, I filter the detectors that are above startingcells/totaldetectors count.
These remaining filtered detectors are then processed to return their median values in their peak detectors.

From here, I take the mean to isolate the bright detectors. So far isolates the stained beads, which is the goal. Not sure how degraded etc handle it, but one thing at a time.