The last of the universal ggplot imports is now gone from the namespace (we will see if this breaks things).
NxN is now using returntype, and has restored name for use in filtering the .csv for gatelines. Replaced base R for filtering with dplyr, as for D17 BCG was causing the entire column to be taken.

Updated UnityPlots pdfs to use FileName instead of AggregateName which didn't exist. Similarly replaced the base R with dplyr call.

We will see if they break.