Wetlab_Decisions is now operational. Iterates through the specimens in Wetlab_Rest data, decides if enough cells are present. If not, uses TheConditions.csv to determine the minimal acceptable cutoffs, and proceeds until the surviving conditions are determined or sample deemed insufficient (and placeholder created).
Each of the individual conditions (either full or best-can-do) calculates out own resuspension volume and then gets bind_row to return the final data.frame for all specimens present.

Room in the parent wrapper to add in the condition treatment arguments if desired, or we can create a new function separate for that.