Luciernaga_QC() is running better for desiredAF arguments. SingleStainSignatures will be where the checking needs to occur next time.
Added compact to RetainedDF to handle the PerCP edge case of when we lost the minor AF signature after subtraction and it no longer was the main signature. This would be where we would use a revised enumeration approach for quantification.

And PerCP-Cy5.5 internal, internal_general, external generate different results as expectedd (wasn't case last night).