Transferred over code for the remaining functions from their respective quarto markdowns. They would just need to be adaptive for the temp folder specific context we are working in.
We are likely going to need to add a .csv panel for the various arguments rather than try to rewrite the existing one, as that would just be more confusing than it is worth.

Of course, a whole day to additionally edit out the bugs would be nice (looking at you Brightness).