Added a bunch of the ggplot2 functions to select functions importFrom to reduce the undefined global function variable burden.
As of now, BiocCheck and check() pass with no errors. Warnings Bioc are missing fields somewhere in the roxygen. Warnings Cran are the spaces in the extdata files, and vignettes starting with number, and Utility_GatingPlots technical debt arguments.

Beyond that 7 and 8 notes respectively. We would want to eliminate warnings entirely pre-submission.