Changed Luciernaga_QC to fix the ugly looking line plot issue.
Swapped RetainedType default to raw at function level.

When RetainedType is normalized, instead of grabbing the ceiling iterated outputs, just sent to AveragedSignature with a normalize=TRUE option that renormalizes

When passed to AveragedSignature with above setting, normalizes and returns a round 3. The function is set to normalize FALSE to avoid unintended consequences.